Reorganize feature spec files
**Why**: One feature per spec file
