+++ spec/features/service_providers_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-feature 'ServiceProviders CRUD' do
+feature 'Service Providers CRUD' do
   context 'Regular user' do
     scenario 'can create service provider' do
       user = create(:user)
@@ -94,70 +94,5 @@ feature 'ServiceProviders CRUD' do
     click_on 'Delete'
 
     expect(page).to have_content('Success')
-  end
-end
-
-feature 'Admin User Approval' do
-  scenario 'only admin user has option to approve service_provider' do
-    user = create(:user)
-    login_as(user)
-
-    visit new_service_provider_path
-
-    expect(page).to_not have_content('Approved')
-  end
-
-  scenario 'admin user has option to approve service_provider' do
-    app = create(:service_provider)
-    admin_user = create(:user, admin: true)
-    login_as(admin_user)
-
-    visit edit_service_provider_path(app)
-
-    expect(page).to have_content('Approved')
-  end
-end
-
-feature 'Users can access sps in their user group' do
-  context 'user is not the creator of the app' do
-    scenario 'service providers from a user group show on index' do
-      group = create(:user_group)
-      user1 = create(:user, user_group: group)
-      user2 = create(:user)
-      user_group_app = create(:service_provider, user_group: group, user: user2)
-      user_created_app = create(:service_provider, user: user1)
-      na_app = create(:service_provider)
-      login_as(user1)
-
-      visit service_providers_path
-
-      expect(page).to have_content(user_group_app.friendly_name)
-      expect(page).to have_content(user_created_app.friendly_name)
-      expect(page).to_not have_content(na_app.friendly_name)
-    end
-
-    scenario 'user can edit a service provider from their user group' do
-      group = create(:user_group)
-      user1 = create(:user, user_group: group)
-      user2 = create(:user)
-      app = create(:service_provider, user_group: group, user: user2)
-      new_name = 'New Service Name'
-      new_description = 'New Description'
-
-      login_as(user1)
-
-      visit edit_service_provider_path(app)
-      fill_in 'Friendly name', with: new_name
-      fill_in 'Description', with: new_description
-      check 'last_name'
-      click_on 'Update'
-
-      expect(page).to have_content('Success')
-      within('table.horizontal-headers') do
-        expect(page).to have_content(new_name)
-        expect(page).to have_content(new_description)
-        expect(page).to have_content('last_name')
-      end
-    end
   end
 end
